allow auth.verifyToken(token, { kid: "specific" })
refs SAS-1471

sometimes we know which key a jwt needs to verify with, even if it's not
in the jwt's header. further, in these cases, we want to fail if the
jwt's header specifies a different key.

while adding this functionality, refactor the remainder. should behave
same as before, but in a more understandable and possibly more efficient
manner.

Change-Id: Iaa543bd6c30c19390f5bdf2887dedfd028b5dce6
Reviewed-on: https://gerrit.instructure.com/c/node-jwt/+/236306
Reviewed-by: Jonathan Featherstone <jfeatherstone@instructure.com>
Tested-by: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>